e2e_tests/python/requirements.txt
aiobotocore[boto3]==2.23.0
boto3==1.38.27
mcp==1.10.1
uvicorn==0.35.0

# For testing, this module is installed from local files.
# Uncomment this line to build using the module from PyPi
# run-mcp-servers-with-aws-lambda >= 0.0.1
